Move hex utils to a subpackage of  `common/`
